Created a builder for the tiles so TileManager is not so large
